Resolve FQDNs in the recipe
I've had issues where performing the DNS resolution when loading the rules
was slow. Plus, some DNS records will return multiple IP addresses, and we
want to add all of them to the firewall (not just one of them).
For these reasons, AFW will perform a full DNS resolution in the recipe, and
the ruleset will only contain IP addresses.